+++ src/app/archbishop/archbishop.py
@@ -22,6 +22,20 @@ def construct_blueprint(bolt, config, messages, redis):
         """ Test request (useful for spinning server up after inactivity) """
         return Response("Hello, World!", status=200)
 
+    @archbishop.route("/scroll", methods=['POST'])
+    def unroll():
+        """ Handle slack command for scroll information """
+        group = redis.get_complex(request.form["team_id"], Group)
+        formatted_scroll = ""
+        position = 0
+        for record in group.scroll:
+            position += 1
+            formatted_scroll += messages.load_with_params(
+                "command.scroll.entry",
+                [str(position), record.name, str(record.streak), str(record.date)]
+            )
+        return Response(formatted_scroll, status=200)
+
     @archbishop.route("/group/<group_id>")
     def group(group_id):
         """ Get group information by id """
@@ -37,7 +51,7 @@ def construct_blueprint(bolt, config, messages, redis):
     @bolt.message(messages.load("event.message.keyword"))
     def handle_waffle(message, say):
         """ Receive and process new waffle score """
-        log.debug(f"[handle_waffle] New Waffle Score received!")
+        log.debug("[handle_waffle] New Waffle Score received!")
         event = Event(message)
         group = get_group(event)
         player = get_player(event, group)
@@ -94,7 +108,7 @@ def construct_blueprint(bolt, config, messages, redis):
 
     def process_result(group, player):
         group.update_player(player)
-        # group.update_scroll(player) # FixMe :: Issue #24
+        group.update_scroll(player)
 
         # Player is the King...
         if player.name == group.king.name:
@@ -123,8 +137,8 @@ def construct_blueprint(bolt, config, messages, redis):
                 else:
                     text = messages.load_with_params("result.common.win", [player.name, str(player.score)])
 
-        Result = namedtuple("Result", "group text")
-        return Result(group, text)
+        result = namedtuple("Result", "group text")
+        return result(group, text)
 
     def present(result, to_channel):
         message = {
@@ -138,7 +152,7 @@ def construct_blueprint(bolt, config, messages, redis):
                         "type": "mrkdwn",
                         "text": (
                                 result + " :blush:\n\n"
-                                       "*I am under development*"
+                                         "*I am under development*"
                         ),
                     },
                 }