+++ documentation/CHANGELOG.md
@@ -7,13 +7,17 @@
 - `0.3.x` Scroll Command
 - `0.4.x` Presentation 
 - `0.5.x` Reactions
-- `1.0.0` App Distribution
+- `1.0.x` App Distribution
 
 # Releases
 <!-- @LatestFirst -->
+## [0.3.0]
+[Scroll Command](https://github.com/jrsmth/waffle-bot/milestone/4) (18/03/2024)
+- `#12` Add a slack /command to access scroll information [![user](https://img.shields.io/badge/adamj335-181717.svg?style=flat&logo=github)](https://github.com/adamj335)
+- `#24` Issues with update_scroll (conflicting data types Record vs Munch dict) [![user](https://img.shields.io/badge/adamj335-181717.svg?style=flat&logo=github)](https://github.com/adamj335)
 
 ## [0.2.0]
-[Data Maturity](https://github.com/jrsmth/waffle-bot/milestone/6) (08/03/2024)
+[Data Maturity](https://github.com/jrsmth/waffle-bot/milestone/8) (08/03/2024)
 - `#30` Formalise the conversion of objects into dictionaries (and vice versa) when moving to and from redis [![user](https://img.shields.io/badge/jrsmth-181717.svg?style=flat&logo=github)](https://github.com/jrsmth) [![user](https://img.shields.io/badge/adamj335-181717.svg?style=flat&logo=github)](https://github.com/adamj335)
 - `#35` Player not getting updated (score & streak) [![user](https://img.shields.io/badge/jrsmth-181717.svg?style=flat&logo=github)](https://github.com/jrsmth) [![user](https://img.shields.io/badge/adamj335-181717.svg?style=flat&logo=github)](https://github.com/adamj335)
 
@@ -38,3 +42,4 @@
 [0.1.0]: https://github.com/jrsmth/waffle-bot/compare/0.0.0...0.1.0
 [0.1.1]: https://github.com/jrsmth/waffle-bot/compare/0.1.0...0.1.1
 [0.2.0]: https://github.com/jrsmth/waffle-bot/compare/0.1.1...0.2.0
+[0.3.0]: https://github.com/jrsmth/waffle-bot/compare/0.2.0...0.3.0